test(KUI-1520): some tests for checkIfOngoingRegistration
