+++ server/apiCalls/getFilteredData.js
@@ -13,6 +13,7 @@ const {
   getSemesterForDate,
 } = require('../util/semesterUtils')
 const i18n = require('../../i18n')
+const { checkIfOngoingRegistration } = require('../util/ongoingRegistration')
 const koppsCourseData = require('./koppsCourseData')
 const ladokApi = require('./ladokApi')
 const courseApi = require('./kursinfoApi')
@@ -122,46 +123,6 @@ const createPeriodString = (ladokRound, periods, language) => {
       return periodCodes.length > 0 ? `${periodString}${periodCodes.join(', ')}</p>` : ''
     })
     .join('')
-}
-
-const checkIfOngoingRegistration = (startDate, periods) => {
-  const registrationDates = { HT: '03-15', VT: '09-15', ST: '02-15' }
-  let code = ''
-  let semester = ''
-  let year = ''
-  let registrationDateString = ''
-  const matchedPeriod = periods.find(
-    period =>
-      startDate >= period.Giltighetsperiod.Startdatum &&
-      startDate <= period.Giltighetsperiod.Slutdatum &&
-      ['HT', 'VT'].includes(period.Kod.slice(0, 2))
-  )
-  if (!matchedPeriod) {
-    // Summer
-    semester = 'ST'
-    year = startDate.substring(0, 4)
-    registrationDateString = `${year}-${registrationDates.ST}`
-  } else {
-    code = matchedPeriod?.Kod
-    semester = code?.slice(0, 2)
-    year = code?.slice(2, 6)
-    registrationDateString = `${year}-${registrationDates[semester]}`
-  }
-
-  const date = new Date(registrationDateString)
-  const day = date.getDay()
-
-  if (day === 6) {
-    date.setDate(date.getDate() + 2)
-  } else if (day === 0) {
-    date.setDate(date.getDate() + 1)
-  }
-  const [firstRegistrationDate] = date.toISOString().split('T')
-
-  const currentDate = new Date()
-  const [formattedCurrentDate] = currentDate.toISOString().split('T')
-
-  return formattedCurrentDate >= firstRegistrationDate && formattedCurrentDate < startDate
 }
 
 function _getRound(koppsRoundObject = {}, ladokRound, socialSchedules, periods, language = 'sv') {
@@ -293,8 +254,6 @@ const getFilteredData = async ({ courseCode, language, memoList }) => {
     ladokApi.getPeriods(),
     getSocial(courseCode, language),
   ])
-
-  console.log(`LADOK ROUNDS: ${JSON.stringify(ladokRounds, null, 4)}`)
 
   //* **** Course information that is static on the course side *****//
   const courseDefaultInformation = _parseCourseDefaultInformation(ladokCourse, ladokSyllabus, language)
